wip. lil refactor of GuttersProps, delete some opportunity unused files, changed the query, moving the settings dialog out of the Visualizer component
