+++ src/domain/collaboration/InnovationFlow/InnovationFlowVisualizers/InnovationFlowChips.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button, Divider, styled, Tooltip } from '@mui/material';
 import { ArrowRight } from '@mui/icons-material';
@@ -9,15 +9,9 @@ import TranslationKey from '../../../../core/i18n/utils/TranslationKey';
 import i18n from '../../../../core/i18n/config';
 import { InnovationFlowState } from '../InnovationFlow';
 import { Caption } from '../../../../core/ui/typography';
+import { InnovationFlowVisualizerProps } from './InnovationFlowVisualizer';
 
-interface InnovationFlowChipsProps {
-  states: InnovationFlowState[];
-  currentState?: string;
-  selectedState: string | undefined;
-  settings?: ReactNode;
-  createButton?: ReactNode;
-  onSelectState?: (state: InnovationFlowState) => void;
-}
+interface InnovationFlowChipsProps extends InnovationFlowVisualizerProps {}
 
 const FlowStateDescription = styled(WrapperMarkdown)(() => ({
   img: {
@@ -32,7 +26,7 @@ const InnovationFlowChips = ({
   currentState,
   selectedState,
   onSelectState,
-  settings,
+  settingsButton,
   createButton,
 }: InnovationFlowChipsProps) => {
   const { t } = useTranslation();
@@ -103,7 +97,7 @@ const InnovationFlowChips = ({
             </Button>
           ))}
         </Gutters>
-        {settings}
+        {settingsButton}
         {createButton && (
           <Gutters row disablePadding>
             <Divider orientation="vertical" flexItem />
