+++ src/domain/collaboration/InnovationFlow/InnovationFlowVisualizers/InnovationFlowVisualizer.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react';
+import { InnovationFlowState } from '../InnovationFlow';
+
+export interface InnovationFlowVisualizerProps {
+  states: InnovationFlowState[];
+  currentState?: string;
+  selectedState: string | undefined;
+  onSelectState?: (state: InnovationFlowState) => void;
+  createButton?: ReactNode;
+  settingsButton?: ReactNode;
+}
